requirements.txt
crewai>=1.0.0
crewai-tools>=0.14.0
requests>=2.31.0
groq
python-dotenv
pypdf
fastapi
uvicorn
python-multipart
pydantic
pyyaml
openai
appdirs
